feat: display `SafeMessage` message hash when signing off-chain
