+++ src/components/transactions/TxDetails/Summary/SafeTxHashDataRow/index.tsx
@@ -1,53 +1,8 @@
-import { TypedDataEncoder } from 'ethers'
 import { TxDataRow, generateDataRowValue } from '../TxDataRow'
 import { type SafeTransactionData, type SafeVersion } from '@safe-global/safe-core-sdk-types'
-import { getEip712TxTypes } from '@safe-global/protocol-kit/dist/src/utils'
 import useSafeAddress from '@/hooks/useSafeAddress'
 import useChainId from '@/hooks/useChainId'
-import semverSatisfies from 'semver/functions/satisfies'
-
-const NEW_DOMAIN_TYPE_HASH_VERSION = '>=1.3.0'
-
-export function getDomainHash({
-  chainId,
-  safeAddress,
-  safeVersion,
-}: {
-  chainId: string
-  safeAddress: string
-  safeVersion: SafeVersion
-}): string {
-  const includeChainId = semverSatisfies(safeVersion, NEW_DOMAIN_TYPE_HASH_VERSION)
-  return TypedDataEncoder.hashDomain({
-    ...(includeChainId && { chainId }),
-    verifyingContract: safeAddress,
-  })
-}
-
-const NEW_SAFE_TX_TYPE_HASH_VERSION = '>=1.0.0'
-
-export function getMessageHash({
-  safeVersion,
-  safeTxData,
-}: {
-  safeVersion: SafeVersion
-  safeTxData: SafeTransactionData
-}): string {
-  const usesBaseGas = semverSatisfies(safeVersion, NEW_SAFE_TX_TYPE_HASH_VERSION)
-  const SafeTx = getEip712TxTypes(safeVersion).SafeTx
-
-  // Clone to not modify the original
-  const tx: any = { ...safeTxData }
-
-  if (!usesBaseGas) {
-    tx.dataGas = tx.baseGas
-    delete tx.baseGas
-
-    SafeTx[5].name = 'dataGas'
-  }
-
-  return TypedDataEncoder.hashStruct('SafeTx', { SafeTx }, tx)
-}
+import { getDomainHash, getSafeTxMessageHash } from '@/utils/safe-hashes'
 
 export const SafeTxHashDataRow = ({
   safeTxHash,
@@ -62,7 +17,7 @@ export const SafeTxHashDataRow = ({
   const safeAddress = useSafeAddress()
 
   const domainHash = getDomainHash({ chainId, safeAddress, safeVersion })
-  const messageHash = safeTxData ? getMessageHash({ safeVersion, safeTxData }) : undefined
+  const messageHash = safeTxData ? getSafeTxMessageHash({ safeVersion, safeTxData }) : undefined
 
   return (
     <>
